Karl: Changed BigDecimal to BigInteger

+++ src/net/azib/java/students/t654321/hometasks/lecture1/Fibonacci.java
@@ -4,8 +4,7 @@
  */
 package net.azib.java.students.t654321.hometasks.lecture1;
 
-import java.math.BigDecimal;
-import java.util.Iterator;
+import java.math.BigInteger;
 
 /**
  *
@@ -14,9 +13,9 @@ import java.util.Iterator;
 public class Fibonacci {
 
     public static void calculateFibonacciElement(int numberOfElement) {
-        BigDecimal[] fibonacciElement = new BigDecimal[numberOfElement];
-        fibonacciElement[0] = BigDecimal.ZERO;
-        fibonacciElement[1] = BigDecimal.ONE;
+        BigInteger[] fibonacciElement = new BigInteger[numberOfElement];
+        fibonacciElement[0] = BigInteger.ZERO;
+        fibonacciElement[1] = BigInteger.ONE;
         for (int i = 2; i < fibonacciElement.length; i++) {
             fibonacciElement[i] = fibonacciElement[i - 2].add(fibonacciElement[i - 1]);
             if (i == numberOfElement - 1) {
